cleaned objects for axios requests in login/sign up forms, moved this pages to Components folder
