+++ water-my-plants/src/Components/SignupForm.js
@@ -7,13 +7,17 @@ const initialState = {
     username:"",
     password:"",
     phone_number:"",
+};
+
+const initialErrorState = {
     error_msg:"",
     error_visible: false
-};
+}
 
 export default function SignupForm(props) {
 
-    const [ newUser, setNewUser] = useState(initialState);
+    const [newUser, setNewUser] = useState(initialState);
+    const [errorData, setErrorData] = useState(initialErrorState);
     const history = useHistory();
 
     const changeHandler = (event) => {
@@ -31,31 +35,31 @@ export default function SignupForm(props) {
         let phoneNumberLength = newUser.phone_number.length;
 
         if (usernameLength < 4) {
-            setNewUser({
-                ...newUser,
+            setErrorData({
+                ...errorData,
                 error_visible:true,
                 error_msg:"username must be at least 4 charachers"
-            });
+            })
         }
         else if (passwordLength < 6) {
-            setNewUser({
-                ...newUser,
+            setErrorData({
+                ...errorData,
                 error_visible:true,
                 error_msg:"password must be at least 6 characters"
-            });
+            })
         }
         else if (phoneNumberLength !== 10) {
-            setNewUser({
-                ...newUser,
+            setErrorData({
+                ...errorData,
                 error_visible:true,
                 error_msg:"invalid phone number"
             });
         }
         else {
-            setNewUser({
-                ...newUser,
+            setErrorData({
+                ...errorData,
                 error_visible:false,
-                error_msg: ""
+                error_msg:""
             });
             console.log("success: ", newUser);
         }
@@ -106,8 +110,8 @@ export default function SignupForm(props) {
                         onChange={changeHandler}
                     />
                 </label>
-                {newUser.error_visible ? <ErrorMsgDiv>
-            <ErrorMsgP>{newUser.error_msg}</ErrorMsgP>
+                {errorData.error_visible ? <ErrorMsgDiv>
+            <ErrorMsgP>{errorData.error_msg}</ErrorMsgP>
             </ErrorMsgDiv> : null}
             </FormDiv>
             <SubmitButton type='submit'>Submit!</SubmitButton>
